Added BigDecimal for better accuracy.

+++ src/models.rs
@@ -1,6 +1,7 @@
 use crate::schema::{locations, readings};
 use chrono::{DateTime, Local};
 use serde::Deserialize;
+use bigdecimal::BigDecimal;
 
 #[derive(Deserialize, Insertable)]
 pub struct Reading {
@@ -8,7 +9,7 @@ pub struct Reading {
     pub id: i32,
     pub index: i32,
     pub field_description: String,
-    pub measurement: f32,
+    pub measurement: BigDecimal,
 }
 
 #[derive(Deserialize, Insertable)]
@@ -16,6 +17,6 @@ pub struct Location {
     pub publication_time: DateTime<Local>,
     pub id: i32,
     pub name: String,
-    pub latitude: f32,
-    pub longitude: f32,
+    pub latitude: BigDecimal,
+    pub longitude: BigDecimal,
 }
